requirements.txt
streamlit
vaderSentiment
keybert
sentence-transformers
langchain
openai
python-dotenv
pyyaml
